requirements.txt
Flask==3.0.0
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
Werkzeug==3.0.1
easyocr==1.7.0
Pillow==10.1.0
WTForms==3.1.1
email-validator==2.1.0
pytest==7.4.3
pytest-cov==4.1.0
reportlab==4.0.7
openpyxl==3.1.2
numpy==1.24.3

